Lesson 9
Task 1
Improve the task with figures
Use the Sealed modifier for the "Figure" class.

Task 2
Complete the problem from homework number 5.
Multiplying two matrices
Create two 3x3 arrays of integers (two matrices).
Write a program to multiply two matrices.
First array: {{1,0,0,0},{0,1,0,0},{0,0,0,0}}
Second array: {{1,2,3},{1,1,1},{0,0,0},{2,1,0}}
Expected result: 1 2 3 1 1 1 0 0 0
Create a Utils class.
In this class, create a static method to print the matrix on the screen using Syste.out.println.
The method must take a two-dimensional array as input, and it must not return anything as output.
Use this static method in a class to solve a matrix multiplication problem.
Use this static method to output a matrix three times - output the first matrix, output the second matrix, output the result.

Task 3
Create a program for transferring from one client’s bank card to his other card.
The data for translation is specified from the console.
There is a client.
The system stores information about the client’s first name, last name, passport number, and date of birth.
As well as his postal code, country, city, street and house number.
The client may have a bank card (one or more).
Make a method to transfer an amount from one multiple to another.
Each card contains the card number, CVC/CVV, current amount on the card, card code(id).
Cards can be of three types: BelCard, MasterCard, VisaCard.
There are limits on transfers. Limits are checked only for the card from which the transfer is made.
Each type of card has its own limits.
Cards may be in different currencies - therefore, provide for conversion.
A commission is charged for the transfer (set as a percentage).
Provide for various checks and validations.
Scenario to test:
The code(id) of the card from... is entered from the console (for example, 1)
the code (2) of the card to... is entered from the console (for example, 2)
The transfer amount is entered from the console.
As a result of the program running, the following message is displayed:
Client's First Name and Last Name
Number of customer cards
Card 1 - amount before transfer
Card 1 - amount after transfer
Card 2 - amount before transfer
Card 2 - amount after transfer
The transfer fee.